move css files into styles folder
