+++ src/components/weather.component.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./weather.style.css";
+import "../styles/weather.style.css";
 
 const Weather = (props) => {
   return (
