--- src/form.style.css
@@ -1,12 +1,0 @@
-form {
-  margin-top: 3rem;
-}
-
-form input[name="city"],
-form input[name="country"] {
-  background-color: transparent;
-  color: white;
-  border: none;
-  border-bottom: 2px solid whitesmoke;
-  border-radius: 0;
-}
